# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Fridge AI Backend API
  version: 1.0.0
  description: OpenAPI spec generated from current Spring controllers.
servers:
  - url: http://localhost:8080
    description: Local backend
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: Get all users
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserList'
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
...
  /api/v1/users/me:
    get:
      tags: [Current User]
      summary: Get current user
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
    put:
      tags: [Current User]
      summary: Update current user profile
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
  /api/v1/inventory:
    get:
      tags: [Inventory App]
      summary: Get current user inventory items
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItemDto'
    post:
      tags: [Inventory App]
      summary: Create inventory item
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItemRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemDto'
  /api/v1/inventory/{id}:
    put:
      tags: [Inventory App]
      summary: Update inventory item
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItemRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemDto'
    delete:
      tags: [Inventory App]
      summary: Delete inventory item
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Success
...
  /api/v1/analytics/events:
    post:
      tags: [Analytics]
      summary: Create analytics event
      parameters:
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsEventRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAnalyticsEvent'
  /api/v1/analytics/events/user/{userId}:
    get:
      tags: [Analytics]
      summary: Get analytics events by user
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAnalyticsEventList'
  /api/v1/analytics/events/type/{eventType}:
    get:
      tags: [Analytics]
      summary: Get analytics events by type
      parameters:
        - name: eventType
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAnalyticsEventList'
  /api/v1/analytics/events/user/{userId}/type/{eventType}:
    get:
      tags: [Analytics]
      summary: Get analytics events by user and type
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - name: eventType
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAnalyticsEventList'
  /api/v1/household/overview:
    get:
      tags: [Household]
      summary: Get household overview
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdOverviewDto'
  /api/v1/household/invite/regenerate:
    post:
      tags: [Household]
      summary: Regenerate invite
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteDto'
  /api/v1/household/members/{memberId}/transfer-owner:
    post:
      tags: [Household]
      summary: Transfer household ownership
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
        - $ref: '#/components/parameters/MemberIdPath'
      responses:
        '200':
          description: Success
  /api/v1/household/members/{memberId}:
    delete:
      tags: [Household]
      summary: Remove household member
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
        - $ref: '#/components/parameters/MemberIdPath'
      responses:
        '200':
          description: Success
  /api/v1/household/leave:
    post:
      tags: [Household]
      summary: Leave household
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '200':
          description: Success
components:
  parameters:
    AuthorizationHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    XUserIdHeader:
      name: X-User-Id
      in: header
      required: true
      schema:
        type: integer
        format: int64
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    UserIdPath:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    MemberIdPath:
      name: memberId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required: [email, password, name]
    RefreshRequest:
      type: object
      properties:
        token:
          type: string
      required: [token]
    UserDto:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserDto'
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]
    UserRequest:
      type: object
      additionalProperties: true
    UserResponse:
      type: object
      additionalProperties: true
...
    AnalyticsEventRequest:
      type: object
      additionalProperties: true
    AnalyticsEventResponse:
      type: object
      additionalProperties: true
    InventoryItemRequest:
      type: object
      additionalProperties: true
    InventoryItemDto:
      type: object
      additionalProperties: true
    HouseholdOverviewDto:
      type: object
      additionalProperties: true
    InviteDto:
      type: object
      additionalProperties: true
    ApiResponseUser:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserResponse'
    ApiResponseUserList:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
...
    ApiResponseAnalyticsEvent:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/AnalyticsEventResponse'
    ApiResponseAnalyticsEventList:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsEventResponse'